Always pass -L to curl
As a rule, always pass `-L` to `curl` to have it follow redirects.  The
big motivation for this is that we now redirect from plain http:// to
https:// in a buncha scenarios, which means a lot of the `curl` examples
are wrong.  And adding `-L` shouldn't ever break what we're demonstrating;
so I just added it to all of the commands (even those that already say
https://) to avoid thinking about it too much and spending more time on
it.

And while I'm at it, fix “curly quotes” being used instead of "stright
quotes" in a command example in knative.md; since the change had me
looking at that line anyway.  (Likely introduced by macOS automatically
doing curly quotes by default...)